native-camera/pyproject.toml
[tool.poetry]
name = "native-camera"
version = "0.1.0"
description = ""
authors = ["Evelyn ?.", "Dylan S."]
readme = "README.md"
packages = [{include = "native_camera"}]

[tool.poetry.dependencies]
python = "^3.8"
python-socketio = "^5.7.2"
aiortc = "^1.3.2"
requests = "^2.28.1"
websocket-client = "^1.4.2"
loguru = "^0.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
